requirements.txt
# ===== 小红书达人爬虫项目依赖 =====

# Playwright: 浏览器自动化工具，用于模拟真实浏览器访问小红书
playwright

# PyYAML: 解析 YAML 配置文件
pyyaml

# SQLAlchemy: ORM 数据库工具，操作 SQLite 数据库
sqlalchemy

# Pydantic: 数据校验与序列化模型
pydantic

# Pandas: 数据处理与分析
pandas

# Openpyxl: 读写 Excel (.xlsx) 文件
openpyxl

# Streamlit: Web 可视化面板框架
streamlit

# Loguru: 简洁强大的日志库
loguru

# HTTPX: HTTP 客户端（用于下载图片）
httpx

# Plotly: 交互式图表（雷达图等）
plotly
